wip(gdn): e2e WORKING — prefill correct + ~1.3x speedup (bf16 fix + managed scratch); decode pending
DIAGNOSTICS (all confirmed):
- Cross-impl A/B (new example gdn_fla_vs_fi): Atlas FLA == FlashInfer on identical input (cos=0.999993,
  norm_ratio=1.0) => GDN math equivalent; the e2e garbage was NOT a math bug.
- ROOT CAUSE = dtype: export/harness were fp16, Atlas GDN q/k/v/o are BF16 -> fp16 kernel misread bf16
  bits. FIX: re-export bf16 (gdn_export.py torch.bfloat16) + relink libatlasgdn.so.
- Also fixed async use-after-free: managed shim entry atlas_gdn_prefill_packed_managed caches
  tensormaps/init/cu internally; gdn_flashinfer.rs calls it (no per-call alloc/free/sync).

E2E (holo35b, ATLAS_GDN_FLASHINFER 1 vs 0, same binary):
- prefill CORRECT (first token matches FLA).
- prefill SPEEDUP ~1.30x @2K, 1.34x @4K, up to 1.46x @C8 (11x is the chunk_delta_h sub-kernel; e2e
  Amdahl-bound by GDN's ~24-38% prefill share, grows with context).
- decode drifts = known state k<->v transpose (FI writes S[v][k], decode reads S[k][v]) — next fix.

+++ 3rdparty_patches/gdn_aot/gdn_export.py
@@ -13,7 +13,7 @@ dr.cached_compile = _wrap
 
 from flashinfer.gdn_prefill import chunk_gated_delta_rule
 T, Hqk, Hv, D = 2048, 16, 32, 128
-dt, dev = torch.float16, "cuda"
+dt, dev = torch.bfloat16, "cuda"
 q = torch.randn(T, Hqk, D, dtype=dt, device=dev)
 k = torch.nn.functional.normalize(torch.randn(T, Hqk, D, dtype=torch.float32, device=dev), dim=-1).to(dt)
 v = torch.randn(T, Hv, D, dtype=dt, device=dev)
